backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
Pillow==10.3.0
django-filter==24.2
python-decouple==3.8
reportlab==4.2.0
openpyxl==3.1.2
celery==5.4.0
redis==5.0.4
django-redis==5.4.0
gunicorn==22.0.0
whitenoise==6.6.0